+++ src/main/java/net/minecraft/block/BlockPotato.java
@@ -20,7 +20,7 @@ public class BlockPotato extends BlockCrops {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, fortune);
 
         if (!worldIn.isRemote) {
-            if (((Integer) state.getValue(AGE)).intValue() >= 7 && worldIn.rand.nextInt(50) == 0) {
+            if (state.getValue(AGE).intValue() >= 7 && worldIn.rand.nextInt(50) == 0) {
                 spawnAsEntity(worldIn, pos, new ItemStack(Items.poisonous_potato));
             }
         }
