+++ src/main/java/net/minecraft/client/entity/EntityOtherPlayerMP.java
@@ -34,8 +34,8 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer {
         this.otherPlayerMPX = x;
         this.otherPlayerMPY = y;
         this.otherPlayerMPZ = z;
-        this.otherPlayerMPYaw = (double) yaw;
-        this.otherPlayerMPPitch = (double) pitch;
+        this.otherPlayerMPYaw = yaw;
+        this.otherPlayerMPPitch = pitch;
         this.otherPlayerMPPosRotationIncrements = posRotationIncrements;
     }
