+++ src/main/java/net/minecraft/command/CommandReplaceItem.java
@@ -139,7 +139,7 @@ public class CommandReplaceItem extends CommandBase {
         if (!SHORTCUTS.containsKey(shortcut)) {
             throw new CommandException("commands.generic.parameter.invalid", new Object[]{shortcut});
         } else {
-            return ((Integer) SHORTCUTS.get(shortcut)).intValue();
+            return SHORTCUTS.get(shortcut).intValue();
         }
     }
